+++ TransitionAnimation/PageTransitionAnimation.swift
@@ -1,16 +1,14 @@
 //
-//  DefaultPushTransitionAnimation.swift
+//  PageTransitionAnimation.swift
 //  TransitionTreasury
 //
-//  Created by DianQK on 16/1/24.
-//  Copyright © 2016年 TransitionTreasury. All rights reserved.
+//  Created by DianQK on 12/30/15.
+//  Copyright © 2016 TransitionTreasury. All rights reserved.
 //
 
-#if TR_MODULE
 import TransitionTreasury
-#endif
-/// Apple Default Push Transition
-public class DefaultPushTransitionAnimation: NSObject, TRViewControllerAnimatedTransitioning, TransitionInteractiveable {
+/// Page Motion
+public class PageTransitionAnimation: NSObject, TRViewControllerAnimatedTransitioning, TransitionInteractiveable {
     
     public var transitionStatus: TransitionStatus
     
@@ -19,15 +17,22 @@ public class DefaultPushTransitionAnimation: NSObject, TRViewControllerAnimatedT
     public var percentTransition: UIPercentDrivenInteractiveTransition?
     
     public var completion: (() -> Void)?
-    
+
     public var cancelPop: Bool = false
-    
+
     public var interacting: Bool = false
     
+    private var transformBackup: CATransform3D?
     private var shadowOpacityBackup: Float?
     private var shadowOffsetBackup: CGSize?
     private var shadowRadiusBackup: CGFloat?
     private var shadowPathBackup: CGPath?
+    
+    private lazy var maskView: UIView = {
+        let maskView = UIView(frame: UIScreen.mainScreen().bounds)
+        maskView.backgroundColor = UIColor.blackColor()
+        return maskView
+    }()
     
     public init(status: TransitionStatus = .Push) {
         transitionStatus = status
@@ -35,7 +40,7 @@ public class DefaultPushTransitionAnimation: NSObject, TRViewControllerAnimatedT
     }
     
     public func transitionDuration(transitionContext: UIViewControllerContextTransitioning?) -> NSTimeInterval {
-        return 0.3
+        return 0.6
     }
     
     public func animateTransition(transitionContext: UIViewControllerContextTransitioning) {
@@ -44,40 +49,52 @@ public class DefaultPushTransitionAnimation: NSObject, TRViewControllerAnimatedT
         var toVC = transitionContext.viewControllerForKey(UITransitionContextToViewControllerKey)
         let containView = transitionContext.containerView()
         
-        var fromVCStartX: CGFloat = 0
-        var fromVCEndX = -UIScreen.mainScreen().bounds.width/3
+        var startPositionX: CGFloat = UIScreen.mainScreen().bounds.width
+        var endPositionX: CGFloat = 0
         
-        var toVCStartX = UIScreen.mainScreen().bounds.width
-        var toVCEndX: CGFloat = 0
+        var startOpacity: Float = 0
+        var endOpacity: Float = 0.3
+        
+        transformBackup = transformBackup ?? fromVC?.view.layer.transform
+        
+        var transform3D: CATransform3D = CATransform3DIdentity
+        transform3D.m34 = -1.0/500.0
         
         if transitionStatus == .Pop {
             swap(&fromVC, &toVC)
-            swap(&fromVCStartX, &fromVCEndX)
-            swap(&toVCStartX, &toVCEndX)
+            swap(&startPositionX, &endPositionX)
+            swap(&startOpacity, &endOpacity)
+        } else {
+            transform3D = CATransform3DTranslate(transform3D, 0, 0, -35)
         }
         
         containView?.addSubview(fromVC!.view)
         containView?.addSubview(toVC!.view)
+        fromVC?.view.addSubview(maskView)
         
-        fromVC?.view.frame.origin.x = fromVCStartX
-        
-        toVC?.view.frame.origin.x = toVCStartX
+        maskView.layer.opacity = startOpacity
+        toVC?.view.layer.position.x = startPositionX + toVC!.view.layer.bounds.width / 2
         shadowOpacityBackup = toVC?.view.layer.shadowOpacity
         shadowOffsetBackup = toVC?.view.layer.shadowOffset
         shadowRadiusBackup = toVC?.view.layer.shadowRadius
         shadowPathBackup = toVC?.view.layer.shadowPath
-        toVC?.view.layer.shadowOpacity = 0.3
+        toVC?.view.layer.shadowOpacity = 0.5
         toVC?.view.layer.shadowOffset = CGSize(width: -3, height: 0)
         toVC?.view.layer.shadowRadius = 5
         toVC?.view.layer.shadowPath = CGPathCreateWithRect(toVC!.view.layer.bounds, nil)
         
         UIView.animateWithDuration(transitionDuration(transitionContext), delay: 0, options: .CurveEaseInOut, animations: {
-            fromVC?.view.frame.origin.x = fromVCEndX
-            toVC?.view.frame.origin.x = toVCEndX
+            self.maskView.layer.opacity = endOpacity
+            fromVC?.view.layer.transform = transform3D
+            toVC?.view.layer.position.x = endPositionX + toVC!.view.layer.bounds.width / 2
             }) { (finished) -> Void in
                 transitionContext.completeTransition(!transitionContext.transitionWasCancelled())
                 if !self.cancelPop {
                     toVC?.view.layer.shadowOpacity = 0
+                    if self.transitionStatus == .Pop && finished && !self.cancelPop {
+                        self.maskView.removeFromSuperview()
+                        fromVC?.view.layer.transform = self.transformBackup ?? CATransform3DIdentity
+                    }
                     if finished {
                         toVC?.view.layer.shadowOpacity = self.shadowOpacityBackup ?? 0
                         toVC?.view.layer.shadowOffset = self.shadowOffsetBackup ?? CGSize(width: 0, height: 0)