+++ TransitionAnimation/TRTransitionAnimations.swift
@@ -5,9 +5,8 @@
 //  Created by 宋宋 on 16/2/5.
 //  Copyright © 2016年 TransitionTreasury. All rights reserved.
 //
-#if TR_MODULE
+
 import TransitionTreasury
-#endif
 /**
  ViewController Push Methods